refactor(factory-sessions): move invocation routing out of service.go
The orchestrator-kind routing pushed ondemandtarget/service.go to 1058
lines, past the repository's 1000-line file limit, taking backend-size
from 98 violations to 99. Moving it to its own file restores the baseline
without weakening the limit or the routing.

No behavior change.
